refactor: consolidate issue status routers into single issue.status.ts
- Merge issueStatus.ts (CRUD) and issue.status.ts (counts) into single router
- Follow consistent issue.*.ts naming pattern used throughout codebase
- Update root.ts import path
- Delete duplicate issueStatus.ts file
- Maintain all functionality: getAll, create, update, delete, getStatusCounts
